Extract FrameArray into its own file
