+++ lib/frames.js
@@ -1,4 +1,6 @@
 "use strict";
+var frame_array_1 = require("./frames/frame-array");
+exports.FrameArray = frame_array_1.FrameArray;
 var frame_arg_1 = require("./frames/frame-arg");
 exports.FrameArg = frame_arg_1.FrameArg;
 var frame_char_1 = require("./frames/frame-char");
@@ -15,4 +17,3 @@ var frame_symbol_1 = require("./frames/frame-symbol");
 exports.FrameSymbol = frame_symbol_1.FrameSymbol;
 var frame_1 = require("./frames/frame");
 exports.Frame = frame_1.Frame;
-exports.FrameArray = frame_1.FrameArray;
