+++ lib/frames/frame-array.d.ts
@@ -1,0 +1,8 @@
+import { Frame } from "./frame";
+export declare class FrameArray extends Frame {
+    protected data: Array<Frame>;
+    constructor(data: Array<Frame>, meta?: {
+        [key: string]: Frame;
+    });
+    at(index: number): Frame;
+}
