+++ lib/frames/frame-expr.js
@@ -6,6 +6,7 @@ var __extends = (this && this.__extends) || function (d, b) {
 };
 var frame_1 = require("./frame");
 var frame_arg_1 = require("./frame-arg");
+var frame_array_1 = require("./frame-array");
 var frame_name_1 = require("./frame-name");
 var FrameExpr = (function (_super) {
     __extends(FrameExpr, _super);
@@ -40,6 +41,6 @@ var FrameExpr = (function (_super) {
     FrameExpr.EXPR_BEGIN = "(";
     FrameExpr.EXPR_END = ")";
     return FrameExpr;
-}(frame_1.FrameArray));
+}(frame_array_1.FrameArray));
 exports.FrameExpr = FrameExpr;
 ;
