+++ lib/frames/frame.js
@@ -1,9 +1,4 @@
 "use strict";
-var __extends = (this && this.__extends) || function (d, b) {
-    for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p];
-    function __() { this.constructor = d; }
-    d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
-};
 exports.Void = {};
 var Frame = (function () {
     function Frame(meta) {
@@ -93,16 +88,3 @@ var Frame = (function () {
 }());
 exports.Frame = Frame;
 ;
-var FrameArray = (function (_super) {
-    __extends(FrameArray, _super);
-    function FrameArray(data, meta) {
-        if (meta === void 0) { meta = exports.Void; }
-        _super.call(this, meta);
-        this.data = data;
-    }
-    FrameArray.prototype.at = function (index) {
-        return this.data[index];
-    };
-    return FrameArray;
-}(Frame));
-exports.FrameArray = FrameArray;
