+++ src/frames.ts
@@ -1,3 +1,4 @@
+export { FrameArray } from "./frames/frame-array";
 export { FrameArg } from "./frames/frame-arg";
 export { FrameChar } from "./frames/frame-char";
 export { FrameLazy } from "./frames/frame-lazy";
@@ -5,4 +6,4 @@ export { FrameExpr } from "./frames/frame-expr";
 export { FrameName } from "./frames/frame-name";
 export { FrameString } from "./frames/frame-string";
 export { FrameSymbol } from "./frames/frame-symbol";
-export { IKeyValuePair, Frame, FrameArray } from "./frames/frame";
+export { IKeyValuePair, Frame } from "./frames/frame";
