+++ src/frames/frame-array.ts
@@ -1,0 +1,11 @@
+import { Frame, Void } from "./frame";
+
+export class FrameArray extends Frame {
+  constructor(protected data: Array<Frame>, meta = Void) {
+    super(meta);
+  }
+
+  public at(index: number) {
+    return this.data[index];
+  }
+}
